Corrections model

+++ src/gof/model/Enseignement.java
@@ -39,7 +39,7 @@ public class Enseignement extends ElemStruct {
 	
 	@Column(length = 2000)
 	protected String discipline;
-	protected int dureeStage;
+	protected float dureeStage;
 	protected String etatRof;
 	protected int langue;
 	@Column(length = 2000)
@@ -81,7 +81,7 @@ public class Enseignement extends ElemStruct {
 			String bibliographie, int capacite, String capitalisation,
 			String coefficient, String competences, String competencesHab,
 			String contenu, String contenuHab, Date dateModification,
-			String discipline, int dureeStage, String etatRof, int langue,
+			String discipline, float dureeStage, String etatRof, int langue,
 			String mcc, String mccHab, String modalitesOrganisation,
 			boolean mutualisable, String preRequis, String preRequisOblig,
 			String preRequisObligHab, String typeEns, int version,
@@ -209,11 +209,11 @@ public class Enseignement extends ElemStruct {
 		this.discipline = discipline;
 	}
 
-	public int getDureeStage() {
+	public float getDureeStage() {
 		return dureeStage;
 	}
 
-	public void setDureeStage(int dureeStage) {
+	public void setDureeStage(float dureeStage) {
 		this.dureeStage = dureeStage;
 	}
